% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_data}
\alias{clean_data}
\title{Clean Emissions Data file}
\usage{
clean_data(type, data_folder)
}
\arguments{
\item{type}{A character string of energy types(e.g. "Coal", "Natural_gas", etc.)}

\item{data_folder}{Path to the directory containing the downloaded Excel 
file.}
}
\value{
A cleaned data frame for the given energy type.
}
\description{
Performs basic cleaning of the Excel file: 
- rename headers 
- remove unwanted rows/columns
- Converts all column and state names to lowercase
}
\examples{
# Provide a path to the folder containing the file, not the file itself.
# clean_data("Coal", path/to/data/folder/with/name/Emission_data)

}
