% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{co2_constants}
\alias{co2_constants}
\title{Constants For Quick Use}
\description{
These constants define default values used throughout the package for 
user input validation, in cases of incorrect, incomplete or blank inputs. 
Some are exported for user access while others are set to internal.
}
\keyword{internal}
