% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_check}
\alias{data_check}
\title{Simple Function for Data Integrity Check}
\usage{
data_check(dataframe)
}
\arguments{
\item{dataframe}{A data frame after filtering to confirm data availability}
}
\value{
Boolean return based on data availability in data frame. TRUE if data 
frame is empty, otherwise FALSE
}
\description{
This function will check if there is any data within the filtered 
data based on number of rows available within a data frame.
}
\examples{
# filtered_list <- process_energy_data(c("Coal"), "tx", 2000, 2010)
# data_check(filtered_list[[1]])
}
\keyword{internal}
