% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_EXT_VALUES}
\alias{DEFAULT_EXT_VALUES}
\alias{DEFAULT_TYPE}
\alias{DEFAULT_USER_STATE}
\alias{DEFAULT_START_YEAR}
\alias{DEFAULT_END_YEAR}
\alias{DEFAULT_OUTPUT}
\title{Default Exported Values For Data}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 1.
}
\usage{
DEFAULT_TYPE

DEFAULT_USER_STATE

DEFAULT_START_YEAR

DEFAULT_END_YEAR

DEFAULT_OUTPUT
}
\description{
These constants define default values used across the package
when the users enters incorrect values, incomplete, or blank inputs.
}
\keyword{datasets}
