% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_INT_VALUES}
\alias{DEFAULT_INT_VALUES}
\alias{DEFAULT_EMISSION_FILENAME}
\alias{DEFAULT_GAS}
\alias{DEFAULT_MIN}
\alias{DEFAULT_MAX}
\alias{PLOT_WIDTH}
\alias{PLOT_HEIGHT}
\alias{PLOT_DPI}
\alias{DEFAULT_VISUAL_OUTPUT}
\title{Default Internal Values For Logic & Validation}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 2.
}
\usage{
DEFAULT_EMISSION_FILENAME

DEFAULT_GAS

DEFAULT_MIN

DEFAULT_MAX

PLOT_WIDTH

PLOT_HEIGHT

PLOT_DPI

DEFAULT_VISUAL_OUTPUT
}
\description{
These constants define internal safeguards across the package for user
input validation and logic handling.
}
\keyword{internal}
