% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_energy_file.R
\name{download_energy_file}
\alias{download_energy_file}
\title{Downloading Energy Data File}
\usage{
download_energy_file()
}
\value{
No return value. Excel file is saved to disk and messages are printed to 
inform user of progress.
}
\description{
Downloads Excel(.xlsx) file for CO2 emissions data based on energy types 
selected by the user. Available options include: "Coal", "Natural gas", 
"Petroleum" or "Total".
}
\examples{
# download_energy_file() 

}
