% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ENERGY_OPTIONS}
\alias{ENERGY_OPTIONS}
\title{Energy options allowed by user}
\format{
An object of class \code{character} of length 4.
}
\usage{
ENERGY_OPTIONS
}
\description{
Energy options allowed by user
}
\keyword{internal}
