% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_data}
\alias{extract_data}
\title{Data Extraction From Processed Data}
\usage{
extract_data(processed_data)
}
\arguments{
\item{processed_data}{A named list of filtered data frames, keyed by energy 
type (e.g. "Coal", "Natural_gas", "Petroleum", "Total"). Each data frame 
includes "state", years" and "emissions_produced" columns.}
}
\value{
A list of values for labeling diagram and naming files.
}
\description{
Extraction of information from a data frame in the list for labeling
}
\examples{
# processed_data <- process_energy_data(c("Coal"), "ca", 1980, 2020)
# extract_data(processed_data)

}
