% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{internal_utils}
\alias{internal_utils}
\title{Utility File Containing Functions Used Throughout The Package}
\description{
This file contains utility functions that support core package features, 
including data validation, transformation, messaging and extraction logic.
}
\keyword{internal}
