% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{is_blank_check}
\alias{is_blank_check}
\title{Simple Function to Check Inputs are Valid}
\usage{
is_blank_check(input)
}
\arguments{
\item{input}{An input of any type for checking input validity}
}
\value{
Boolean return based on validity of input. TRUE if input is any of 
blank, NA or NULL, otherwise FALSE.
}
\description{
This function will check if there is any blank, NA or NULL values in the 
inputs for the function to work independently.
}
\examples{
# is_blank_check("Coal")  # Returns FALSE
# is_blank_check("")  # Returns TRUE
}
\keyword{internal}
