% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_energy_data.R
\name{process_energy_data}
\alias{process_energy_data}
\title{Emission Data Loading, Cleaning & Filtering}
\usage{
process_energy_data(energy_type, us_state, start_year, end_year)
}
\arguments{
\item{energy_type}{A character vector of one or more energy types (e.g. 
c("Coal", "Natural gas", "Petroleum", "Total") or any variation of this 
format.}

\item{us_state}{A character string specifying the US state or whole US to 
filter by.}

\item{start_year}{A numerical value indicating the start of the 
filtering range. Defaults to first available value in the data if NA, NULL 
or empty when co2emissionanalyzer() is used first.}

\item{end_year}{A numerical value indicating the end of the 
filtering range. Defaults to first available value in the data if NA, NULL 
or empty when co2emissionanalyzer() is used first.}
}
\value{
A named list of filtered data frames, keyed by energy type.
}
\description{
Loads, cleans and filters emissions data for one or more energy type(s)
and selected state. Data is cleaned by removing unwanted rows and columns, 
and transforming to long format for easier analysis and plotting. Includes 
checks for function to operate independently. Defaults set when running
co2emissionanalyzer().
}
\examples{
# process_energy_data(c("Coal", "Petroleum", "Total"), "ca", 1980, 2020)

}
