% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2emissionsanalyzer.R
\name{start_energy_analysis}
\alias{start_energy_analysis}
\title{Start the full energy analysis pipeline}
\usage{
start_energy_analysis()
}
\value{
None. Side effects include user interactions, data downloading and 
plotting generation.
}
\description{
This function guides the user through downloading and analyzing of CO2 
emissions data by energy resource type and U.S. state. It requires no 
initial input, but user is interactively prompted to provide  the necessary
parameters for data processing and manipulation.
}
\details{
The emissions data is sourced from U.S. Energy Information Administration at
this link: \url{https://www.eia.gov/state/seds/sep_sum/html/xls/co2_source.xlsx}
}
\examples{
# start_energy_analysis()

}
