% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_STATES}
\alias{VALID_STATES}
\title{List of 50 US States & District of Columbia}
\format{
An object of class \code{character} of length 132.
}
\usage{
VALID_STATES
}
\description{
List of 50 US States & District of Columbia
}
\keyword{internal}
