% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_energy_data.R
\name{visualize_energy_data}
\alias{visualize_energy_data}
\title{Visualization of CO2 Emissions by Energy Type}
\usage{
visualize_energy_data(processed_data, output)
}
\arguments{
\item{processed_data}{A named list of filtered data frames, keyed by energy 
type (e.g. "Coal", "Natural_gas", "Petroleum", "Total"). Each data frame 
includes "state", "years" and "emissions_produced" columns.}

\item{output}{A string specifying the desired output type. Options 
are "line" (for a graphical line diagram) or "table" (for an Excel output)
Default is "line".}
}
\value{
If output = "line", displays completion message and print 
last ggplot created. If output = "table", saves an Excel file and displays 
completion message to user. In both cases, **invisible NULL is returned**.
}
\description{
Graphically representing CO2 emissions by the different or all energy 
resources based on user requirements. Users can generate a line diagram 
saved as an image, or an Excel file with tabular emissions data for 
each energy type. This function is based on the previous output from
process_energy_data.
}
\examples{
# Before running, make sure to download the data:
# download_energy_file()

# Then process the data:
# processed <- process_energy_data(c("Coal", "Total"), "ca", 1980, 2020)

# Finally visualize it:
# visualize_energy_data(processed, "line")

}
