% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Simple Function To Change Frame Format}
\usage{
wide_to_long(dataframe)
}
\arguments{
\item{dataframe}{A data frame to transform from wide to long format through 
pivot}
}
\value{
A long-format data frame with numeric 'years' and a single 
'emissions_produced' column.
}
\description{
Converts a data frame from wide to long format, and converts year values 
from character to numeric.
}
\examples{
# df <- data.frame(state = c("ca", "tx"),
#                `2010` = c(100, 200),
#                `2011` = c(110, 210))
# wide_to_long(df)
}
